modify torch_util
